Added OrderMapper

+++ src/main/java/com/example/clothingstore/model/Product.java
@@ -47,7 +47,7 @@ public class Product {
         this.description = description;
     }
 
-    public BigDecimal getPrice() {
+    public double getPrice() {
         return price;
     }
